Add a param to not fetch recursively submodules
Add a `disable_recursive_submodules` parameter, if `true` the `in` resourse will not fetch submodules recursively.

We sometimes doesn't need to fetch all submodules, if we don't fetch them we can gain times during task.

Usecase:
We have a bosh release like diego-release which has script and manifests to generate a final manifest, we only need those files and set as submodules this release.
But when using this resource we will fetch all its submodules which is useless.